Fix for size_t in UpnpString. (cherry picked from commit ee5bd670d4eeb09bf2ebc0e01a902ebac93aa5b8)

+++ upnp/src/api/UpnpString.c
@@ -49,7 +49,7 @@ extern char *strndup(__const char *__string, size_t __n);
 struct SUpnpString
 {
 	/*! \brief Length of the string. */
-	int m_length;
+	size_t m_length;
 	/*! \brief Pointer to a dynamically allocated area that holds the NULL
 	 * terminated string. */
 	char *m_string;
